move toastr options into unique utilities file
