+++ src/components/usuarios/UsuariosListItem.jsx
@@ -6,7 +6,9 @@ import { toastr } from 'react-redux-toastr';
 
 import axios from 'axios';
 
-import { CloudFunctionsUrl, ToastrOptionsError, ToastrConfirmOptions, ToastrOptionsSuccess } from '../../utilities/constants.js';
+import { CloudFunctionsUrl } from '../../utilities/constants.js'
+
+import { ToastrOptionsError, ToastrOptionsConfirm, ToastrOptionsSuccess } from '../../utilities/toastr.js';
 
 const UsuariosListItem = ({ usuario, firebase, currentUID }) => {
   const { key } = usuario;
@@ -83,7 +85,7 @@ const UsuariosListItem = ({ usuario, firebase, currentUID }) => {
         <div className='grid-item'>
           <button
             className={key === currentUID ? '' : 'u-pointer font-bold'}
-            onClick={() => toastr.confirm('¿Seguro que deseas eliminar esta usuario?', ToastrConfirmOptions(removeUser, key))}
+            onClick={() => toastr.confirm('¿Seguro que deseas eliminar esta usuario?', ToastrOptionsConfirm(removeUser, key))}
             disabled={key === currentUID}
           >Eliminar</button>
         </div>
